refactor(api): extract locale-resolution + schema-loading + languages domains (#35)
